split out follow tests by entity type
